Refaktoroitu haku-listauksen suodattimet ja taulukko erillisiksi reiteiksi
