+++ src/app/(root)/@controls/loading.tsx
@@ -1,0 +1,5 @@
+import { FullSpinner } from '@/app/components/full-spinner';
+
+export default function Loading() {
+  return <FullSpinner />;
+}
