+++ src/app/(root)/loading.tsx
@@ -1,0 +1,5 @@
+import { FullSpinner } from '../components/full-spinner';
+
+export default function Loading() {
+  return <FullSpinner />;
+}
